Longに変更

+++ src/main/java/com/example/macaron/SuddenSale.java
@@ -2,13 +2,13 @@ package com.example.macaron;
 
 //completedのみ編集可能
 public class SuddenSale {
-    private final int id;
+    private final Long id;
     private final int store_id;
     private final String name;
     private final String contents;
     private boolean completed;
 
-    public SuddenSale(int id, int store_id, String name, String contents, boolean completed) {
+    public SuddenSale(Long id, int store_id, String name, String contents, boolean completed) {
         this.id = id;
         this.store_id = store_id;
         this.name = name;
@@ -16,7 +16,7 @@ public class SuddenSale {
         this.completed = completed;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
     public int getStore_id() {
